v4l2/stateless: Frame id normalization
The v4l2_buffer->timestamp field is the source of truth. |timestamp|
is a struct timeval. The member |tv_usec| is used to hold the frame id
and is of type uint64_t.

Cq-Include-Trybots: luci.chromium.try:linux-v4l2-codec-rel
Bug: b:278935312
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5147350
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240360}

+++ media/gpu/v4l2/stateless/queue.cc
@@ -170,7 +170,7 @@ void InputQueue::Reclaim(Buffer& buffer) {
 bool InputQueue::SubmitCompressedFrameData(void* ctrls,
                                            const void* data,
                                            size_t length,
-                                           uint32_t frame_id) {
+                                           uint64_t frame_id) {
   // Failing to acquire a buffer is a normal part of the process. All of the
   // input buffers can be full if the output buffers are not being cleared.
   auto buffer_index = GetFreeBufferIndex();
