+++ media/gpu/v4l2/stateless/stateless_decode_surface.h
@@ -17,7 +17,7 @@ namespace media {
 
 class StatelessDecodeSurface : public base::RefCounted<StatelessDecodeSurface> {
  public:
-  StatelessDecodeSurface(uint32_t frame_id, base::OnceClosure enqueue_cb);
+  StatelessDecodeSurface(uint64_t frame_id, base::OnceClosure enqueue_cb);
 
   StatelessDecodeSurface(const StatelessDecodeSurface&) = delete;
   StatelessDecodeSurface& operator=(const StatelessDecodeSurface&) = delete;
@@ -26,7 +26,7 @@ class StatelessDecodeSurface : public base::RefCounted<StatelessDecodeSurface> {
   void SetColorSpace(const VideoColorSpace& color_space);
   void SetVideoFrameTimestamp(const base::TimeDelta timestamp);
 
-  int32_t FrameID() const { return frame_id_; }
+  uint64_t FrameID() const { return frame_id_; }
   VideoColorSpace ColorSpace() const { return color_space_; }
   base::TimeDelta VideoFrameTimestamp() const { return video_frame_timestamp_; }
   uint64_t GetReferenceTimestamp() const;
@@ -42,7 +42,7 @@ class StatelessDecodeSurface : public base::RefCounted<StatelessDecodeSurface> {
  private:
   // Identify this surface so that it can be matched up the the uncompressed
   // buffer when it is done being decompressed.
-  const uint32_t frame_id_;
+  const uint64_t frame_id_;
 
   // The visible size of the buffer.
   gfx::Rect visible_rect_;
